+++ app/src/androidTest/java/com/sidgowda/pawcalc/navigation/NavigationTest.kt
@@ -7,8 +7,6 @@ import androidx.test.espresso.IdlingRegistry
 import androidx.test.espresso.NoActivityResumedException
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import com.sidgowda.pawcalc.test.fakes.FakeOnboardingDataSourceSingleton
-import com.sidgowda.pawcalc.PawCalcActivity
 import com.sidgowda.pawcalc.R
 import com.sidgowda.pawcalc.data.modules.OnboardingDataModule
 import com.sidgowda.pawcalc.data.onboarding.model.OnboardingState
@@ -18,6 +16,8 @@ import com.sidgowda.pawcalc.db.di.DbModule
 import com.sidgowda.pawcalc.navigation.*
 import com.sidgowda.pawcalc.test.IdlingResourceCoroutineDispatcher
 import com.sidgowda.pawcalc.test.TestTags
+import com.sidgowda.pawcalc.test.fakes.FakeOnboardingDataSourceSingleton
+import com.sidgowda.pawcalc.ui.PawCalcActivity
 import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
