+++ app/src/main/java/com/sidgowda/pawcalc/ui/PawCalcActivity.kt
@@ -1,4 +1,4 @@
-package com.sidgowda.pawcalc
+package com.sidgowda.pawcalc.ui
 
 import android.os.Bundle
 import androidx.activity.compose.setContent
@@ -11,7 +11,10 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
+import com.sidgowda.pawcalc.PawCalcHost
 import com.sidgowda.pawcalc.common.settings.ThemeFormat
+import com.sidgowda.pawcalc.data.onboarding.model.OnboardingState
+import com.sidgowda.pawcalc.data.settings.model.Settings
 import com.sidgowda.pawcalc.ui.theme.Green500
 import com.sidgowda.pawcalc.ui.theme.Grey900
 import com.sidgowda.pawcalc.ui.theme.LightDarkPreview
