+++ app/src/test/java/com/sidgowda/pawcalc/ui/PawCalcViewModelTest.kt
@@ -1,4 +1,4 @@
-package com.sidgowda.pawcalc
+package com.sidgowda.pawcalc.ui
 
 import com.sidgowda.pawcalc.common.settings.DateFormat
 import com.sidgowda.pawcalc.common.settings.ThemeFormat
